upload success

+++ src/main/java/com/file/upload/dto/FileMetadataDto.java
@@ -7,6 +7,6 @@ public class FileMetadataDto {
 	private String accountId;
 	private String fileName;
 	private String fileHash;
-	private Double size;
+	private Float size;
 }
